requirements.txt
# packaging
pip>=9
setuptools>=26
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
#gcsfs
#google-cloud-storage
#mlflow
#s3fs

# utilities
six>=1.14
#joblib
#memoized-property
termcolor
notebook

# dataviz
geopandas==0.10.2
matplotlib
seaborn
plotly==5.8.2
dash==2.5.1
dash_bootstrap_components
plotly-express==0.4.1
python-dotenv
gunicorn
sqlalchemy
requests
pymysql
whitenoise
